Implements fetching and mapping of now playing movies
Fetches now playing movies from The Movie Database API and maps the results to a standardized Movie interface.  This involves creating new interfaces for movies and API responses, as well as a mapper class to handle the transformation.  The movie data is now returned as an array of `Movie` objects.